Update VRT reference images.
