+++ assets/sass/components/analytics-4/_index.scss
@@ -31,6 +31,6 @@
 @import "audience-segmentation/googlesitekit-audience-tiles";
 @import "audience-segmentation/googlesitekit-no-audience-banner";
 @import "audience-segmentation/googlesitekit-audience-widget-area";
-@import "audience-segmentation/googlesitekit-partial-data-badge";
+@import "audience-segmentation/googlesitekit-badge-with-tooltip";
 @import "audience-segmentation/googlesitekit-partial-data-notice";
 @import "audience-segmentation/googlesitekit-audience-selection-panel";
