+++ assets/sass/components/analytics-4/audience-segmentation/_googlesitekit-badge-with-tooltip.scss
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-.googlesitekit-audience-segmentation-partial-data-badge {
+.googlesitekit-badge-with-tooltip {
 	background-color: $c-utility-warning-container;
 	border-radius: $br-lg;
 	color: $c-yellow-y-600;
